wip tags instead of score for signalement
